rating

+++ src/main/java/com/example/pet_care_booking/dto/ProductDTO.java
@@ -33,5 +33,5 @@ public class ProductDTO {
    private String updatedAt;
    private List<ImagesDTO> imagesDTO;
    private Long categoryId;
-   private BigDecimal averageRating;
+   private Double averageRating;
 }
