+++ src/main/java/com/example/pet_care_booking/dto/RatingDTO.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @Data
 @Builder
 public class RatingDTO {
-   private Integer ratingValue;
+   private Double ratingValue;
    private Long id;
    private String comment;
    private LocalDateTime createdAt;
